backend/pyproject.toml
[project]
name = "stockxpert-backend"
version = "0.1.0"
description = "FastAPI backend for StockXpert dashboard"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.0",
    "torch>=2.0.0",
    "scikit-learn>=1.2.0",
    "joblib>=1.3.0",
    "pyyaml>=6.0",
    "pytz>=2024.1",
]
